SECRETARIA-36 - Gestión dinámica de los roles y permisos asignados a usuarios
> Permitir eliminar roles en la vista de edición.
> No se puede eliminar un rol que está asignado a algunos usuarios

+++ src/main/java/es/tenerife/secretaria/libro/web/rest/RolResource.java
@@ -25,6 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.codahale.metrics.annotation.Timed;
 
 import es.tenerife.secretaria.libro.domain.Rol;
+import es.tenerife.secretaria.libro.repository.UsuarioRepository;
 import es.tenerife.secretaria.libro.service.RolService;
 import es.tenerife.secretaria.libro.web.rest.dto.RolDTO;
 import es.tenerife.secretaria.libro.web.rest.mapper.RolMapper;
@@ -43,10 +44,12 @@ public class RolResource extends AbstractResource {
 
 	private RolService rolService;
 	private RolMapper rolMapper;
+	private UsuarioRepository usuarioRepository;
 
-	public RolResource(RolService rolService, RolMapper rolMapper) {
+	public RolResource(RolService rolService, RolMapper rolMapper, UsuarioRepository usuarioRepository) {
 		this.rolService = rolService;
 		this.rolMapper = rolMapper;
+		this.usuarioRepository = usuarioRepository;
 
 	}
 
@@ -121,6 +124,10 @@ public class RolResource extends AbstractResource {
 	@PreAuthorize("hasPermission('ROL', 'ELIMINAR')")
 	public ResponseEntity<RolDTO> deleteRol(@PathVariable String nombre) {
 		log.debug("REST request to get Rol : {}", nombre);
+		if (!usuarioRepository.findAllByRoles_Nombre(nombre).isEmpty()) {
+			return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "usuario-rol",
+					"No se puede eliminar un rol que está asignado a algunos usuarios")).body(null);
+		}
 		rolService.delete(nombre);
 		return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, nombre)).build();
 	}
